# openapi.yaml
openapi: 3.0.3
info:
  title: Config Management Service
  version: "1.0"
  description: This service handle create, update, and listing configuration
servers:
  - url: http://localhost:3001/api/v1/config
paths:
...
  /:schema/versions:
    get:
      summary: Retrieve all history versions config of the schema
      parameters:
        - name: schema
          in: path
          required: true
          description: Schema Name
          schema:
            type: string
            example: "payment_config"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Success"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentConfig"
        "400":
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      status:
                        type: string
                        example: "Fail"
                      message:
                        type: string
                        example: "Schema not exists"
                  - type: object
                    properties:
                      status:
                        type: string
                        example: "Fail"
                      message:
                        type: string
                        example: "Configuration for payment_config schema not found"
components:
  schemas:
    PaymentConfig:
      type: object
      properties:
        max_limit:
          type: int
          example: 1000
        enabled:
          type: boolean
          example: false
...